+++ packages/controls/js/libs/repeater-control/components/actions.js
@@ -82,6 +82,10 @@ export default function RepeaterItemActions({
 	} = useControlContext();
 
 	function closeMenu(event: MouseEvent) {
+		if (actionButtonsType !== 'menu') {
+			return;
+		}
+
 		const escEvent = new KeyboardEvent('keydown', {
 			key: 'Escape',
 			code: 'Escape',
@@ -126,6 +130,12 @@ export default function RepeaterItemActions({
 	function deleteFunction(event: MouseEvent) {
 		if (event && event?.hasOwnProperty('stopPropagation')) {
 			event.stopPropagation();
+			if (
+				typeof event.nativeEvent?.stopImmediatePropagation ===
+				'function'
+			) {
+				event.nativeEvent.stopImmediatePropagation();
+			}
 		}
 
 		// Try to open the confirm delete modal if it is not open and shouldConfirmDeleteModal is true
@@ -229,6 +239,12 @@ export default function RepeaterItemActions({
 
 	function cloneFunction(event: MouseEvent) {
 		event.stopPropagation();
+		if (event.preventDefault) {
+			event.preventDefault();
+		}
+		if (typeof event.nativeEvent?.stopImmediatePropagation === 'function') {
+			event.nativeEvent.stopImmediatePropagation();
+		}
 		closeMenu(event);
 
 		const performClone = (): void => {
@@ -436,6 +452,9 @@ export default function RepeaterItemActions({
 								className={controlInnerClassNames('btn-clone')}
 								noBorder={true}
 								icon={<Icon icon="clone" iconSize={20} />}
+								onMouseDown={(event: MouseEvent) => {
+									event.stopPropagation();
+								}}
 								onClick={cloneFunction}
 								aria-label={sprintf(
 									// translators: %s is the repeater item id. It's aria label for cloning repeater item